src/camera/camera.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { databaseProviders } from '../database/database.providers';
import { CameraProviders } from './camera.provider';
import { CameraService } from './camera.service';

describe('CameraService', () => {
  let service: CameraService;

  beforeEach(async () => {
    const module: TestingModule = await Test.createTestingModule({
      providers: [...databaseProviders, ...CameraProviders, CameraService],
    }).compile();

    service = module.get<CameraService>(CameraService);
  });

  it('should be defined', () => {
    expect(service).toBeDefined();
  });
  it('should be creating Camera', async () => {
    expect(service).toBeDefined();
    try {
      await service.create({
        name: 'Test',
        color: '#FFF',
        id: 'test',
      });
    } catch (error) {}
  });

  it('should be creating Camera', async () => {
    expect(service).toBeDefined();
    const data = await service.findAll();
    expect(data.length).toBeGreaterThan(0);
  });

  it('should be Camera', async () => {
    expect(service).toBeDefined();
    const data = await service.findOne({ id: 'test' });
    expect(data).toBeDefined();
  });
});
